# fix: clobbered bp issue in assembly on e2 impl