--- a/go.mod
+++ b/go.mod
@@ -4,7 +4,7 @@ go 1.15
 
 require (
 	github.com/consensys/bavard v0.1.8-0.20210302234946-daa97270b001
-	github.com/consensys/goff v0.3.13-0.20210303152514-315a82768027
+	github.com/consensys/goff v0.3.13-0.20210303164546-ecd387e0d3ae
 	github.com/ethereum/go-ethereum v1.9.25
 	github.com/kilic/bls12-381 v0.0.0-20201226121925-69dacb279461
 	github.com/leanovate/gopter v0.2.9