… .github/workflows/develop.yml
… bls377/fp/element_mul_amd64.s
… bls377/fp/element_ops_amd64.s
… bls377/fr/element_mul_amd64.s
… bls377/fr/element_ops_amd64.s
… bls377/internal/fptower/e2_amd64.s
… bls377/internal/fptower/e2_test.go
… bls381/fp/element_mul_adx_amd64.s
… bls381/fp/element_mul_amd64.s
… bls381/fp/element_ops_amd64.s
… bls381/fr/element_mul_adx_amd64.s
… bls381/fr/element_mul_amd64.s
… bls381/fr/element_ops_amd64.s
… bls381/internal/fptower/e2_amd64.s
… bls381/internal/fptower/e2_test.go
… bn256/fp/element_mul_adx_amd64.s
… bn256/fp/element_mul_amd64.s
… bn256/fp/element_ops_amd64.s
… bn256/fr/element_mul_adx_amd64.s
… bn256/fr/element_mul_amd64.s
… bn256/fr/element_ops_amd64.s
… bn256/internal/fptower/e2_test.go
… bw761/fp/element_mul_adx_amd64.s
… bw761/fp/element_mul_amd64.s
… bw761/fp/element_ops_amd64.s
… bw761/fr/element_mul_adx_amd64.s
… bw761/fr/element_mul_amd64.s
… bw761/fr/element_ops_amd64.s
… go.sum
… internal/asm/amd64/e2.go